nmea-simulator-final-tested/nmea-simulator/requirements.txt
# Core dependencies
pyyaml>=6.0
dataclasses-json>=0.5.7
typing-extensions>=4.0.0

# Networking and async
asyncio-mqtt>=0.11.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

